Extract generic traits into util/network
