+++ Cargo.toml
@@ -42,7 +42,7 @@ ethcore-light = { path = "ethcore/light" }
 ethcore-logger = { path = "logger" }
 ethcore-migrations = { path = "ethcore/migrations" }
 ethcore-miner = { path = "miner" }
-ethcore-network-devp2p = { path = "util/network-devp2p" }
+ethcore-network = { path = "util/network" }
 ethcore-stratum = { path = "stratum" }
 ethcore-transaction = { path = "ethcore/transaction" }
 ethereum-types = "0.2"
