+++ ethcore/light/src/lib.rs
@@ -55,7 +55,7 @@ extern crate log;
 
 extern crate bincode;
 extern crate ethcore_io as io;
-extern crate ethcore_network_devp2p as network;
+extern crate ethcore_network as network;
 extern crate ethcore_bytes as bytes;
 extern crate ethcore_transaction as transaction;
 extern crate ethereum_types;
