+++ parity/main.rs
@@ -53,7 +53,7 @@ extern crate ethcore_light as light;
 extern crate ethcore_logger;
 extern crate ethcore_migrations as migrations;
 extern crate ethcore_miner as miner;
-extern crate ethcore_network_devp2p as network;
+extern crate ethcore_network as network;
 extern crate ethcore_transaction as transaction;
 extern crate ethereum_types;
 extern crate migration as migr;
