+++ util/network-devp2p/src/lib.rs
@@ -19,8 +19,10 @@
 //! Example usage for creating a network service and adding an IO handler:
 //!
 //! ```rust
-//! extern crate ethcore_network_devp2p as net;
+//! extern crate ethcore_network as net;
+//! extern crate ethcore_network_devp2p as devp2p;
 //! use net::*;
+//! use devp2p::*;
 //! use std::sync::Arc;
 //!
 //! struct MyHandler;
@@ -78,6 +80,7 @@ extern crate rlp;
 extern crate bytes;
 extern crate path;
 extern crate ethcore_logger;
+extern crate ethcore_network as network;
 extern crate ipnetwork;
 extern crate keccak_hash as hash;
 extern crate serde;
@@ -100,112 +103,17 @@ mod handshake;
 mod session;
 mod discovery;
 mod service;
-mod error;
 mod node_table;
 mod stats;
 mod ip_utils;
 mod connection_filter;
 
-pub use host::{HostInfo, PeerId, PacketId, ProtocolId, NetworkContext, NetworkIoMessage, NetworkConfiguration};
 pub use service::NetworkService;
-pub use error::{Error, ErrorKind};
 pub use stats::NetworkStats;
-pub use session::SessionInfo;
 pub use connection_filter::{ConnectionFilter, ConnectionDirection};
+pub use host::NetworkContext;
 
 pub use io::TimerToken;
 pub use node_table::{validate_node_url, NodeId};
-use ipnetwork::{IpNetwork, IpNetworkError};
-use std::str::FromStr;
 
 const PROTOCOL_VERSION: u32 = 5;
-
-/// Network IO protocol handler. This needs to be implemented for each new subprotocol.
-/// All the handler function are called from within IO event loop.
-/// `Message` is the type for message data.
-pub trait NetworkProtocolHandler: Sync + Send {
-	/// Initialize the handler
-	fn initialize(&self, _io: &NetworkContext, _host_info: &HostInfo) {}
-	/// Called when new network packet received.
-	fn read(&self, io: &NetworkContext, peer: &PeerId, packet_id: u8, data: &[u8]);
-	/// Called when new peer is connected. Only called when peer supports the same protocol.
-	fn connected(&self, io: &NetworkContext, peer: &PeerId);
-	/// Called when a previously connected peer disconnects.
-	fn disconnected(&self, io: &NetworkContext, peer: &PeerId);
-	/// Timer function called after a timeout created with `NetworkContext::timeout`.
-	fn timeout(&self, _io: &NetworkContext, _timer: TimerToken) {}
-}
-
-/// Non-reserved peer modes.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum NonReservedPeerMode {
-	/// Accept them. This is the default.
-	Accept,
-	/// Deny them.
-	Deny,
-}
-
-impl NonReservedPeerMode {
-	/// Attempt to parse the peer mode from a string.
-	pub fn parse(s: &str) -> Option<Self> {
-		match s {
-			"accept" => Some(NonReservedPeerMode::Accept),
-			"deny" => Some(NonReservedPeerMode::Deny),
-			_ => None,
-		}
-	}
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-#[cfg_attr(feature = "ipc", binary)]
-pub struct IpFilter {
-    pub predefined: AllowIP,
-    pub custom_allow: Vec<IpNetwork>,
-    pub custom_block: Vec<IpNetwork>,
-}
-
-impl Default for IpFilter {
-    fn default() -> Self {
-        IpFilter {
-            predefined: AllowIP::All,
-            custom_allow: vec![],
-            custom_block: vec![],
-        }
-    }
-}
-
-impl IpFilter {
-    /// Attempt to parse the peer mode from a string.
-    pub fn parse(s: &str) -> Result<IpFilter, IpNetworkError> {
-        let mut filter = IpFilter::default();
-        for f in s.split_whitespace() {
-            match f {
-                "all" => filter.predefined = AllowIP::All,
-                "private" => filter.predefined = AllowIP::Private,
-                "public" => filter.predefined = AllowIP::Public,
-                "none" => filter.predefined = AllowIP::None,
-                custom => {
-                    if custom.starts_with("-") {
-                        filter.custom_block.push(IpNetwork::from_str(&custom.to_owned().split_off(1))?)
-                    } else {
-                        filter.custom_allow.push(IpNetwork::from_str(custom)?)
-                    }
-                }
-            }
-        }
-        Ok(filter)
-    }
-}
-
-/// IP fiter
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum AllowIP {
-	/// Connect to any address
-	All,
-	/// Connect to private network only
-	Private,
-	/// Connect to public network only
-	Public,
-    /// Block all addresses
-    None,
-}
