+++ util/network-devp2p/src/node_table.rs
@@ -23,8 +23,7 @@ use std::str::FromStr;
 use std::{fs, mem, slice};
 use ethereum_types::H512;
 use rlp::*;
-use error::{Error, ErrorKind};
-use {AllowIP, IpFilter};
+use network::{Error, ErrorKind, AllowIP, IpFilter};
 use discovery::{TableUpdates, NodeEntry};
 use ip_utils::*;
 use serde_json;
