+++ util/network-devp2p/src/session.rs
@@ -16,7 +16,6 @@
 
 use std::{str, io};
 use std::net::SocketAddr;
-use std::cmp::Ordering;
 use std::sync::*;
 use std::collections::HashMap;
 
@@ -28,7 +27,8 @@ use rlp::*;
 use connection::{EncryptedConnection, Packet, Connection, MAX_PAYLOAD_SIZE};
 use handshake::Handshake;
 use io::{IoContext, StreamToken};
-use error::{Error, ErrorKind, DisconnectReason};
+use network::{Error, ErrorKind, DisconnectReason, SessionInfo, ProtocolId, PeerCapabilityInfo};
+use network::{SessionCapabilityInfo, HostInfo as HostInfoTrait};
 use host::*;
 use node_table::NodeId;
 use stats::NetworkStats;
@@ -88,81 +88,6 @@ pub enum SessionData {
 	},
 	/// Session has more data to be read
 	Continue,
-}
-
-/// Shared session information
-#[derive(Debug, Clone)]
-pub struct SessionInfo {
-	/// Peer public key
-	pub id: Option<NodeId>,
-	/// Peer client ID
-	pub client_version: String,
-	/// Peer RLPx protocol version
-	pub protocol_version: u32,
-	/// Session protocol capabilities
-	pub capabilities: Vec<SessionCapabilityInfo>,
-	/// Peer protocol capabilities
-	pub peer_capabilities: Vec<PeerCapabilityInfo>,
-	/// Peer ping delay in milliseconds
-	pub ping_ms: Option<u64>,
-	/// True if this session was originated by us.
-	pub originated: bool,
-	/// Remote endpoint address of the session
-	pub remote_address: String,
-	/// Local endpoint address of the session
-	pub local_address: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PeerCapabilityInfo {
-	pub protocol: ProtocolId,
-	pub version: u8,
-}
-
-impl Decodable for PeerCapabilityInfo {
-	fn decode(rlp: &UntrustedRlp) -> Result<Self, DecoderError> {
-		let p: Vec<u8> = rlp.val_at(0)?;
-		if p.len() != 3 {
-			return Err(DecoderError::Custom("Invalid subprotocol string length. Should be 3"));
-		}
-		let mut p2: ProtocolId = [0u8; 3];
-		p2.clone_from_slice(&p);
-		Ok(PeerCapabilityInfo {
-			protocol: p2,
-			version: rlp.val_at(1)?
-		})
-	}
-}
-
-impl ToString for PeerCapabilityInfo {
-	fn to_string(&self) -> String {
-		format!("{}/{}", str::from_utf8(&self.protocol[..]).unwrap_or("???"), self.version)
-	}
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct SessionCapabilityInfo {
-	pub protocol: [u8; 3],
-	pub version: u8,
-	pub packet_count: u8,
-	pub id_offset: u8,
-}
-
-impl PartialOrd for SessionCapabilityInfo {
-	fn partial_cmp(&self, other: &SessionCapabilityInfo) -> Option<Ordering> {
-		Some(self.cmp(other))
-	}
-}
-
-impl Ord for SessionCapabilityInfo {
-	fn cmp(&self, b: &SessionCapabilityInfo) -> Ordering {
-		// By protocol id first
-		if self.protocol != b.protocol {
-			return self.protocol.cmp(&b.protocol);
-		}
-		// By version
-		self.version.cmp(&b.version)
-	}
 }
 
 const PACKET_HELLO: u8 = 0x80;
