+++ util/network-devp2p/tests/tests.rs
@@ -20,6 +20,7 @@ extern crate parking_lot;
 extern crate ethcore_bytes;
 extern crate ethcore_io as io;
 extern crate ethcore_logger;
+extern crate ethcore_network;
 extern crate ethcore_network_devp2p;
 extern crate ethkey;
 
@@ -30,6 +31,7 @@ use std::time::*;
 use parking_lot::Mutex;
 use ethcore_bytes::Bytes;
 use ethcore_network::*;
+use ethcore_network_devp2p::*;
 use ethkey::{Random, Generator};
 use io::TimerToken;
