+++ whisper/Cargo.toml
@@ -8,7 +8,7 @@ description = "Whisper Protocol implementation for Parity"
 bitflags = "0.9"
 byteorder = "1.0.0"
 ethereum-types = "0.2"
-ethcore-network-devp2p = { path = "../util/network-devp2p" }
+ethcore-network = { path = "../util/network" }
 ethcrypto = { path = "../ethcrypto" }
 ethkey = { path = "../ethkey" }
 hex = "0.2"
